requirements.txt
python-dotenv==1.0.0
transformers==4.28.1
yfinance==0.2.18
